move constants to the constants directory
