+++ app/accounts/payable/upload-file/normalize-data.ts
@@ -1,67 +1,4 @@
-import { Account, AccountState, AccountType, Currency } from '@/models/Account'
-
-export const initialState: Account = {
-  date: '',
-  authorizationNumber: '',
-  dteNumber: '',
-  currency: Currency.GTQ,
-  iva: 0,
-  issuerName: '',
-  accountType: AccountType.PAYABLE,
-  type: '',
-  recipientName: '',
-  receiverId: '',
-  issuerId: '',
-  serie: '',
-  state: AccountState.VIGENTE,
-  amount: 0,
-  metadata: {
-    petroleum: 0,
-    accommodation: 0,
-    tickets: 0,
-    pressStamp: 0,
-    firefighters: 0,
-    municipalTax: 0,
-    tobacco: 0,
-    alcoholicBeverages: 0,
-    cement: 0,
-    nonAlcoholicBeverages: 0,
-    portFee: 0,
-  },
-}
-
-export const headers = [
-  'Fecha de emisión',
-  'Número de Autorización',
-  'Tipo de DTE ',
-  'Serie',
-  'Número del DTE',
-  'NIT del emisor',
-  'Nombre completo del emisor',
-  'Código de establecimiento',
-  'Nombre del establecimiento',
-  'ID del receptor',
-  'Nombre completo del receptor',
-  'NIT del Certificador',
-  'Nombre completo del Certificador',
-  'Moneda',
-  'Monto ',
-  'Fecha de anulación',
-  'Estado',
-  'Marca de anulado',
-  'IVA ',
-  'Petróleo ',
-  'Turismo Hospedaje ',
-  'Turismo Pasajes ',
-  'Timbre de Prensa ',
-  'Bomberos ',
-  'Tasa Municipal ',
-  'Bebidas alcohólicas ',
-  'Tabaco ',
-  'Cemento ',
-  'Bebidas no Alcohólicas ',
-  'Tarifa Portuaria ',
-]
+import { AccountType } from '@/models/Account'
 
 const noPayableValues = [
   'NIT del Certificador',
