+++ constants/PayableAccountHeader.ts
@@ -1,0 +1,32 @@
+export const payableAccountHeader = [
+  'Fecha de emisión',
+  'Número de Autorización',
+  'Tipo de DTE ',
+  'Serie',
+  'Número del DTE',
+  'NIT del emisor',
+  'Nombre completo del emisor',
+  'Código de establecimiento',
+  'Nombre del establecimiento',
+  'ID del receptor',
+  'Nombre completo del receptor',
+  'NIT del Certificador',
+  'Nombre completo del Certificador',
+  'Moneda',
+  'Monto ',
+  'Fecha de anulación',
+  'Estado',
+  'Marca de anulado',
+  'IVA ',
+  'Petróleo ',
+  'Turismo Hospedaje ',
+  'Turismo Pasajes ',
+  'Timbre de Prensa ',
+  'Bomberos ',
+  'Tasa Municipal ',
+  'Bebidas alcohólicas ',
+  'Tabaco ',
+  'Cemento ',
+  'Bebidas no Alcohólicas ',
+  'Tarifa Portuaria ',
+]
